+++ NEWS.md
@@ -3,6 +3,7 @@
 * Features:
     * Incorporate `alphavantager`, a lightweight API to the [Alpha Vantage financial data provider](https://www.alphavantage.co/).
     * Integrate `Rblpapi`, R interface to "Bloomberg"
+    * Add Google Finance as a source in `tq_get(get = "stock.prices.google")`
 
 ## tidyquant 0.5.3